Revert "day 2"
