+++ index.html
@@ -1,16 +1,12 @@
 <!DOCTYPE html>
 <html>
-
-<head>
+  <head>
     <meta charset="utf-8">
     <meta name="viewport" content="width=device-width,initial-scale=1.0">
-    <script src="https://kit.fontawesome.com/your-code-number.js"></script>
     <title>project-intern</title>
-</head>
-
-<body>
+  </head>
+  <body>
     <div id="app"></div>
     <!-- built files will be auto injected -->
-</body>
-
+  </body>
 </html>
