+++ package.json
@@ -14,16 +14,12 @@
     "build": "node build/build.js"
   },
   "dependencies": {
-    "@melmacaluso/vue-modal": "^2.1.0",
     "bootstrap": "^4.3.1",
     "bootstrap-vue": "^2.0.4",
     "vue": "^2.6.10",
-    "vue-awesome": "^4.0.2",
-    "vue-pose": "^0.5.1",
     "vue-router": "^3.0.1"
   },
   "devDependencies": {
-    "@fortawesome/fontawesome-free": "^5.11.2",
     "autoprefixer": "^7.1.2",
     "babel-core": "^6.22.1",
     "babel-eslint": "^8.2.1",
