+++ src/App.vue
@@ -1,15 +1,12 @@
 <template>
   <div id="app">
-    <!-- <img src="./assets/logo.png"> -->
+    <img src="./assets/logo.png">
     <router-view/>
   </div>
 </template>
 
 <script>
 import "bootstrap/dist/css/bootstrap.css";
-import 'bootstrap-vue/dist/bootstrap-vue.css';
-import '@fortawesome/fontawesome-free/css/all.css'
-import '@fortawesome/fontawesome-free/js/all.js'
 
 export default {
   name: 'App'
