--- src/components/DetailEmp.vue
@@ -1,3 +1,0 @@
-<template>
-    
-</template>
