--- src/components/ExampleComponent.vue
@@ -1,11 +1,0 @@
-<template>
-  <div>
-    <p>I am a component</p>
-  </div>
-</template>
-
-<script>
-export default {
-  name: "ExampleComponent"
-};
-</script>
